🩹 fix: 전체적인 오타 및 api 유저 닉네임, 프로필 사진 실시간 연동 수정

+++ src/api/recipe.api.js
@@ -4,7 +4,7 @@ class RecipeAPI {
   // 레시피 목록 불러오기
   async getRecipes() {
     try {
-      const { data, error } = await supabase.from('recipes').select('*');
+      const { data, error } = await supabase.from('recipes').select('*,users(nickname)');
       if (error) {
         throw Error('레시피 목록 데이터 가져오기 실패');
       }
@@ -17,7 +17,7 @@ class RecipeAPI {
   // 내가 작성한 레시피 목록 불러오기
   async getMyRecipes(userId) {
     try {
-      const { data, error } = await supabase.from('recipes').select('*').eq('user_id', userId);
+      const { data, error } = await supabase.from('recipes').select('*,users(nickname)').eq('user_id', userId);
       if (error) {
         throw Error('레시피 목록 데이터 가져오기 실패');
       }
@@ -30,7 +30,7 @@ class RecipeAPI {
   // 레시피 항목 불러오기
   async getRecipeById(recipeId) {
     try {
-      const { data, error } = await supabase.from('recipes').select('*').eq('id', recipeId);
+      const { data, error } = await supabase.from('recipes').select('*,users(nickname)').eq('id', recipeId);
       if (error) {
         throw Error('레시피 항목 데이터 가져오기 실패');
       }
@@ -46,7 +46,7 @@ class RecipeAPI {
     // 파일이 있는 경우에만 이미지 업로드 시도
     if (file) {
       // 스토리지에 이미지 저장
-      const { data: uploadData, error: uploadError } = await supabase.storage
+      const { error: uploadError } = await supabase.storage
         .from('images')
         .upload(`recipeimages/${recipe.id}.${file.name.split('.').pop()}`, file, {
           cacheControl: '3600',
@@ -70,14 +70,6 @@ class RecipeAPI {
 
       thumbnailUrl = reloadData.publicUrl;
     }
-    console.log('Inserting recipe with data:', {
-      id: recipe.id,
-      title: recipe.title,
-      user_id: userId,
-      nickname: nickname,
-      content: recipe.content,
-      thumbnail: thumbnailUrl
-    });
     // 레시피 데이터베이스에 추가
     const { data, error } = await supabase.from('recipes').insert({
       id: recipe.id,
@@ -94,39 +86,54 @@ class RecipeAPI {
       console.log('Recipe inserted successfully:', data);
     }
   }
-
   // 레시피 삭제 메서드
-  async DeleteRecipe(recipe) {
-    // 레시피 썸네일 url 선택
-    const { data: thumbnailData } = await supabase.from('recipes').select('thumbnail').eq('id', recipe);
-    // 레시피 테이블 삭제
-    const { data: recipeDeleteData, error: recipeDeleteError } = await supabase
-      .from('recipes')
-      .delete()
-      .eq('id', recipe);
+  async DeleteRecipe(recipeId) {
+    try {
+      // 레시피 썸네일 데이터 가져오기
+      const { data: thumbnailData, error: thumbnailError } = await supabase
+        .from('recipes')
+        .select('thumbnail')
+        .eq('id', recipeId)
+        .single();
 
-    if (recipeDeleteError) {
-      console.error('Error deleting recipe:', recipeDeleteError);
-      return;
-    }
-    if (thumbnailData) {
-      //스토리지 이미지 삭제
-      const { data: imageDeleteData, error: imageDeleteError } = await supabase.storage
-        .from('images')
-        .remove(`recipeimages/${recipe.id}.${thumbnailData[0].thumbnail.split('.').pop()}`);
-
-      if (imageDeleteError) {
-        console.error('Error deleting image:', imageDeleteError);
-        return;
+      if (thumbnailError) {
+        throw new Error('레시피 썸네일 데이터 가져오기 실패');
       }
 
-      console.log('Recipe and image deleted successfully:', recipeDeleteData);
+      // 레시피 테이블에서 삭제
+      const { data: recipeDeleteData, error: recipeDeleteError } = await supabase
+        .from('recipes')
+        .delete()
+        .eq('id', recipeId);
+
+      if (recipeDeleteError) {
+        throw new Error('레시피 삭제 실패');
+      }
+
+      if (thumbnailData && thumbnailData.thumbnail) {
+        // 이미지 파일 이름에서 확장자 추출
+        const extension = thumbnailData.thumbnail.split('.').pop();
+
+        // 스토리지에서 이미지 삭제
+        const { error: imageDeleteError } = await supabase.storage
+          .from('images')
+          .remove(`recipeimages/${recipeId}.${extension}`);
+
+        if (imageDeleteError) {
+          throw new Error('이미지 삭제 실패');
+        }
+
+        console.log('Recipe and image deleted successfully:', recipeDeleteData);
+      } else {
+        console.log('Recipe deleted successfully:', recipeDeleteData);
+      }
+    } catch (error) {
+      console.error('Error deleting recipe:', error);
     }
   }
 
   // 업데이트 메서드
   async UpdateRecipe(recipe, file) {
-    console.log(recipe);
     const { data: existingRecipe, error: existingRecipeError } = await supabase
       .from('recipes')
       .select('thumbnail')
@@ -153,7 +160,7 @@ class RecipeAPI {
       }
 
       // 새로운 파일 업로드
-      const { data: uploadData, error: uploadError } = await supabase.storage
+      const { error: uploadError } = await supabase.storage
         .from('images')
         .upload(`recipeimages/${recipe.id}.${file.name.split('.').pop()}`, file, {
           cacheControl: '3600',
